Refactor into separate modules.
In preparation for adding complexity in the form of Sentinel support and
a few Sentinel-related utility applications, break the existing code
into modules.
